requirements.txt
fastapi
uvicorn
pyyaml
python-multipart
pytest